+++ salt/resource-manager/files/spark-no-policy.sh
@@ -1,0 +1,2 @@
+echo $1
+exit 0