+++ salt/resource-manager/templates/spark-user-group-policy.sh
@@ -1,55 +1,53 @@
 #!/usr/bin/env bash
 # This code implements a user/group based queue placement policy.
 
-MAP={{ map_file }}
-REQUEST=''
-
-KEEP=()
-# Find the requested queue
-# and keep all other the options
-while [[ $# -gt 0 ]]
-do
-  case "${1}" in
-  --queue)
-      REQUEST="${2}"
-      shift 2
-      ;;
-  *)
-      KEEP+=("${1}")
-      shift
-      ;;
-  esac
-done
-
+define_queue ()
+{
 # See who is submitting the request
-USER_NAME=`id -un`
-USER_GROUPS=`id -Gn`
+local REQUEST=$1
+local USER_NAME=`id -un`
+local USER_GROUPS=`id -Gn`
 
+local MAP={{ map_file }}
 if [ ! -f $MAP ]; then
   echo "$MAP file missing"
-  exit 2
+  exit 1
 fi
 
-TARGET=''
+QUEUE=''
 # Read in the map tables
 while read member queue; do
   [[ "$member" =~ ^#.* ]] && continue
   user=`echo "$member" | awk -F':' '{print $1}'`
   group=`echo "$member" | awk -F':' '{print $2}'`
   if [ "X$user" == "X$USER_NAME" ]; then
-    TARGET="$queue"
-  elif [ "X$group" != "X" ]; then
+    QUEUE="$queue"
+  elif [ "X$user" == "X" ] && [ "X$group" != "X" ]; then
     if echo "$USER_GROUPS" | grep -q -w "$group" ; then
-      TARGET="$queue"
+      QUEUE="$queue"
     fi
   fi
-  if [ "X$TARGET" != "X" ] && [ "X$REQUEST" == "X$TARGET" ]; then
-    # if this is the requested queue, then we are done
-    break;
+  if [ "X$QUEUE" != "X" ]; then 
+    if [ "X$REQUEST" == "X$QUEUE" ] || [ "X$REQUEST" == "X" ] ; then
+      # if this is the requested queue or no queue was requested, then we are done
+      break;
+    else 
+      QUEUE=''
+    fi
   fi
 done < $MAP
+}
 
-# Overwrite any queue setting
-KEEP+=("--queue" "$TARGET")
-set -- "${KEEP[@]}"
-echo "-------Starting with ${KEEP[@]}"
+define_queue $1
+if [ "X$QUEUE" != "X" ]; then
+  echo $QUEUE
+  exit 0
+else
+  if [ "X$1" == "X" ]; then
+    echo "Error: No matching rule."
+    exit 2
+  else
+    echo "Error: No matching queue."
+    exit 3
+  fi
+fi